+++ test_reader.h
@@ -4,4 +4,4 @@
 #include <vector>
 
 std::vector<struct test> init_read_test();
-std::string read_test(struct test test, int index);
+std::string read_test(struct test test, size_t index);
